refactor: Rename `lib/throwAuthError.js`
