+++ lib/output-command.js
@@ -4,7 +4,7 @@ const { isEmpty, isObject } = require('lodash');
 const { writeText } = require('@serverless/utils/log');
 const log = require('./log');
 const isAuthenticated = require('./is-authenticated');
-const throwAuthError = require('./throwAuthError');
+const throwAuthError = require('./throw-auth-error');
 const resolveOutputs = require('./resolve-outputs');
 const resolveOutput = require('./resolve-output');
