+++ lib/plugin.js
@@ -23,7 +23,7 @@ const setApiGatewayAccessLogFormat = require('./set-api-gateway-access-log-forma
 const paramCommand = require('./param-command');
 const outputCommand = require('./output-command');
 const isAuthenticated = require('./is-authenticated');
-const throwAuthError = require('./throwAuthError');
+const throwAuthError = require('./throw-auth-error');
 
 const unconditionalCommands = new Set([
   'dashboard',
