+++ lib/throw-auth-error.js
