+++ lib/variables.js
@@ -4,7 +4,7 @@ const _ = require('lodash');
 const ensureString = require('type/string/ensure');
 const resolveOutput = require('./resolve-output');
 const resolveParams = require('./resolve-params');
-const throwAuthError = require('./throwAuthError');
+const throwAuthError = require('./throw-auth-error');
 
 // functions for new way of getting variables
 const getValueFromDashboardParams = (context) => async (paramName) => {
